requirements/base.txt
Django==5.0.4
psycopg[binary]==3.1.18
structlog==24.1.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
